requirements.txt
astroid==1.5.3
certifi==2017.4.17
chardet==3.0.3
dj-database-url==0.4.2
Django==1.11.2
django-braces==1.11.0
django-filter==1.0.4
django-oauth-toolkit==0.12.0
django-rainbowtests==0.6.0
djangorestframework==3.6.3
gunicorn==19.7.1
httpie==0.9.9
idna==2.5
isort==4.2.13
lazy-object-proxy==1.3.1
Markdown==2.6.8
mccabe==0.6.1
oauthlib==2.0.1
psycopg2==2.7.1
Pygments==2.2.0
pylint==1.7.1
pytz==2017.2
requests==2.17.3
six==1.10.0
urllib3==1.21.1
virtualenv==15.1.0
whitenoise==3.3.0
wrapt==1.10.10
channels==1.1.3
